requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python==4.10.0.84
numpy==1.26.4
tensorflow==2.17.0
protobuf==4.25.3
h5py==3.11.0
python-dotenv==1.0.1
requests==2.32.3
